Move the get_kernel_version_list function to the Live OS utilities
The get_kernel_version_list function can be used only with Live OS payload.
